requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==19.3b0
certifi==2021.5.30
cfgv==3.3.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
distlib==0.3.2
Django==3.2.6
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-imagekit==4.0.2
filelock==3.0.12
flake8==3.9.2
identify==2.2.13
idna==3.2
isort==5.9.3
mccabe==0.6.1
nodeenv==1.6.0
pilkit==2.0
Pillow==8.3.1
platformdirs==2.2.0
pre-commit==2.14.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.7.2
